refactor: extracted resolveConflictingFieldTypes
